fix int[] type in tests

+++ main_test.go
@@ -6,8 +6,8 @@ import (
 
 func TestHistorianHysteria(t *testing.T) {
 	t.Run("1 2 3, 3 2 1 = 0", func(t *testing.T) {
-		listA := []int{1, 2, 3}
-		listB := []int{3, 2, 1}
+		listA := []int64{1, 2, 3}
+		listB := []int64{3, 2, 1}
 		result := HistorianHysteria(listA, listB)
 
 		if result != 0 {
@@ -16,8 +16,8 @@ func TestHistorianHysteria(t *testing.T) {
 	})
 
 	t.Run("2 2 3, 3 2 1 = 1", func(t *testing.T) {
-		listA := []int{2, 2, 3}
-		listB := []int{3, 2, 1}
+		listA := []int64{2, 2, 3}
+		listB := []int64{3, 2, 1}
 		result := HistorianHysteria(listA, listB)
 
 		if result != 1 {
@@ -26,8 +26,8 @@ func TestHistorianHysteria(t *testing.T) {
 	})
 
 	t.Run("3 2 2, 1 1 1 = 4", func(t *testing.T) {
-		listA := []int{2, 2, 3}
-		listB := []int{3, 2, 1}
+		listA := []int64{2, 2, 3}
+		listB := []int64{3, 2, 1}
 		result := HistorianHysteria(listA, listB)
 
 		if result != 1 {
